[refactor] 장소 추천 로직을 Port-Adapter 구조로 분리
