+++ src/main/java/com/bready/server/recommendation/port/PlaceRecommendationPort.java
@@ -1,8 +1,20 @@
 package com.bready.server.recommendation.port;
 
+import com.bready.server.plan.domain.PlanCategory;
+import com.bready.server.recommendation.dto.PlaceRecommendationResponse;
+import com.bready.server.trigger.domain.TriggerType;
+
 import java.util.List;
 
 public interface PlaceRecommendationPort {
 
-    List<Long> recommendPlaceCandidates(Long categoryId, Long userId);
+    List<PlaceRecommendationResponse.RecommendationItem> recommendPlaceCandidates(
+            PlanCategory category,
+            TriggerType triggerType,
+            String region,
+            Double latitude,
+            Double longitude,
+            int radius,
+            int limit
+    );
 }
