+++ src/main/java/com/bready/server/recommendation/service/PlaceRecommendationService.java
@@ -2,11 +2,8 @@ package com.bready.server.recommendation.service;
 
 import com.bready.server.global.exception.ApplicationException;
 import com.bready.server.place.domain.PlaceCandidate;
-import com.bready.server.place.domain.PlaceCategoryType;
-import com.bready.server.place.dto.PlaceSearchResponse;
 import com.bready.server.place.exception.PlaceErrorCase;
 import com.bready.server.place.repository.PlaceCandidateRepository;
-import com.bready.server.place.service.PlaceSearchService;
 import com.bready.server.plan.domain.CategoryState;
 import com.bready.server.plan.domain.Plan;
 import com.bready.server.plan.domain.PlanCategory;
@@ -18,6 +15,7 @@ import com.bready.server.plan.repository.PlanRepository;
 import com.bready.server.recommendation.dto.PlaceRecommendationQuery;
 import com.bready.server.recommendation.dto.PlaceRecommendationRequest;
 import com.bready.server.recommendation.dto.PlaceRecommendationResponse;
+import com.bready.server.recommendation.port.PlaceRecommendationPort;
 import com.bready.server.trigger.domain.Trigger;
 import com.bready.server.trigger.domain.TriggerType;
 import com.bready.server.trigger.exception.TriggerErrorCase;
@@ -41,8 +39,8 @@ public class PlaceRecommendationService {
     private final PlanCategoryRepository planCategoryRepository;
     private final TriggerRepository triggerRepository;
     private final PlaceCandidateRepository placeCandidateRepository;
-    private final PlaceSearchService placeSearchService;
     private final CategoryStateRepository categoryStateRepository;
+    private final PlaceRecommendationPort placeRecommendationPort;
 
     @Transactional(readOnly = true)
     public PlaceRecommendationResponse recommendPlaces(Long userId, PlaceRecommendationRequest request, PlaceRecommendationQuery query) {
@@ -76,33 +74,16 @@ public class PlaceRecommendationService {
 
         Coordinate base = resolveBaseCoordinate(trigger.getTriggerType(), category.getId(), query);
 
-        String keyword = buildKeyword(region, trigger.getTriggerType());
-        PlaceCategoryType categoryType = category.getCategoryType();
-
-        List<PlaceSearchResponse> candidates = placeSearchService.search(
-                categoryType,
-                keyword,
-                base.latitude(),
-                base.longitude(),
-                radius
-        );
-
-        if (candidates == null || candidates.isEmpty()) {
-            throw new ApplicationException(PlaceErrorCase.PLACE_NOT_FOUND);
-        }
-
-        List<PlaceRecommendationResponse.RecommendationItem> items = candidates.stream()
-                .limit(limit)
-                .map(p -> new PlaceRecommendationResponse.RecommendationItem(
-                        p.externalId(),
-                        p.name(),
-                        p.address(),
-                        p.latitude(),
-                        p.longitude(),
-                        p.isIndoor(),
-                        buildReason(trigger.getTriggerType())
-                ))
-                .toList();
+        List<PlaceRecommendationResponse.RecommendationItem> items =
+                placeRecommendationPort.recommendPlaceCandidates(
+                        category,
+                        trigger.getTriggerType(),
+                        region,
+                        base.latitude(),
+                        base.longitude(),
+                        radius,
+                        limit
+                );
 
         if (items.isEmpty()) {
             throw new ApplicationException(PlaceErrorCase.PLACE_NOT_FOUND);
@@ -154,36 +135,10 @@ public class PlaceRecommendationService {
         return Math.max(1, Math.min(size, MAX_SIZE));
     }
 
-    private String buildKeyword(String region, TriggerType triggerType) {
-        String base = triggerKeyword(triggerType);
-        if (region == null || region.isBlank()) return base;
-        return region + " " + base;
-    }
-
     private String firstNonBlank(String a, String b) {
         if (a != null && !a.isBlank()) return a;
         if (b != null && !b.isBlank()) return b;
         return null;
-    }
-
-    private String triggerKeyword(TriggerType triggerType) {
-        return switch (triggerType) {
-            case WEATHER_BAD -> "실내";
-            case WAITING_TOO_LONG -> "근처";
-            case PLACE_CLOSED -> "대체";
-            case FATIGUE -> "휴식";
-            case DISTANCE_TOO_FAR -> "근처";
-        };
-    }
-
-    private String buildReason(TriggerType triggerType) {
-        return switch (triggerType) {
-            case WEATHER_BAD -> "날씨 악화 상황을 고려해 실내 이동이 가능한 장소를 추천합니다.";
-            case WAITING_TOO_LONG -> "혼잡을 피하기 위해 주변 대체 후보를 추천합니다.";
-            case PLACE_CLOSED -> "영업 종료 상황을 고려해 주변 대체 장소를 추천합니다.";
-            case FATIGUE -> "휴식이 가능한 성격의 장소를 우선 추천합니다.";
-            case DISTANCE_TOO_FAR -> "이동 부담을 줄이기 위해 더 가까운 후보를 우선 추천합니다.";
-        };
     }
 
     private record Coordinate(Double latitude, Double longitude) {}
